+++ ANTLR/out.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ "$(basename "$PWD")" != "ANTLR" ]; then
+    cd "ANTLR" || exit 1
+fi
+
+./check_lib.sh
+
 rm -rf ./Java
 mkdir ./Java
 
@@ -15,7 +21,7 @@ java -jar /usr/local/lib/antlr-4.13.1-complete.jar ./Imports.g4 -o ./Java -visit
 ls ./Java/*.java
 echo "Imports File Java files generated"
 
-javac ./Java/$TYPE*.java
+javac -cp /usr/local/lib/antlr-4.13.1-complete.jar ./Java/$TYPE*.java
 ls ./Java/
 echo "Java files compiled"
